+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+
 ## [1.0.2] - 2022-01-03
 ### Changed
 - Include additional information with error status codes.
@@ -28,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial release.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.0.2..HEAD
 [1.0.2]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.0.1..v1.0.2
 [1.0.1]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.0.0..v1.0.1
 [1.0.0]: https://github.com/DeepLcom/deepl-dotnet/compare/v0.1.0...v1.0.0