+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+* Fix issue in .NET versions earlier than 5.0, when creating large glossaries or
+ translating texts larger than 64 KiB.
+### Security
+
+
 ## [1.0.3] - 2022-01-19
 ### Added
 * Add contribution guidelines -- currently we are unable to accept Pull Requests.
@@ -35,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial release.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.0.3..HEAD
 [1.0.3]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.0.2..v1.0.3
 [1.0.2]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.0.1..v1.0.2
 [1.0.1]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.0.0..v1.0.1