+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+
 ## [1.1.0] - 2022-04-13
 ### Added
 * Add `ErrorMessage` property to `DocumentStatus` that contains a short
@@ -56,6 +65,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial release.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.1.0..HEAD
 [1.1.0]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.0.5..v1.1.0
 [1.0.5]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.0.4..v1.0.5
 [1.0.4]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.0.3..v1.0.4