+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Changed
+* Update contributing guidelines, we can now accept Pull Requests.
+
+
 ## [1.2.0] - 2022-05-18
 ### Added
 * New languages available: Indonesian (`'id'`) and Turkish (`'tr'`). Add
@@ -71,6 +76,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial release.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.2.0..HEAD
 [1.2.0]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.1.0..v1.2.0
 [1.1.0]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.0.5..v1.1.0
 [1.0.5]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.0.4..v1.0.5