+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 * Update contributing guidelines, we can now accept Pull Requests.
+### Fixed
+* Fix bug in OutlineDetection option for text translation.
 
 
 ## [1.2.0] - 2022-05-18